// apps/customer-site/src/components/layout/footer/types.ts
export type SocialLinks = {
  tiktok?: string;
  facebook?: string;
  instagram?: string;
  twitter?: string;
  youtube?: string;
};

export type FooterWebsiteSettings = {
  company_name: string;
  logoUrl: string;
  social_links: SocialLinks;
  address?: string;
  phone?: string;
  email?: string;
};

export type FooterDictionary = any;